requirements.txt
Bio==1.8.1
biopython==1.85
Markdown==3.7
neo4j==5.28.1
pandas==2.3.3
plotnine==0.15.0
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic-ai-slim==1.42.0 
pydantic-ai-slim[google]
pydantic_core==2.41.5
pyodbc==5.2.0
python-dotenv==1.2.1
Requests==2.32.5
fastapi>=0.115.0
uvicorn==0.24.0
fastapi[standard]